app/face_recognition/requirements2.txt
click==8.1.7
cmake==3.30.5
colorama==0.4.6
cv==1.0.0
cv2box==0.5.9
dlib @ file:///D:/WORKSPACE/BTL-PYTHON/dlib-19.22.99-cp310-cp310-win_amd64.whl#sha256=f2181b6724669fb0147c9ebc326d72b94de41bced1812eff5df4b133b5d0b575
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.26.4
opencv-python==4.10.0.84
pandas==2.2.3
pillow==11.0.0
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
tqdm==4.66.5
tzdata==2024.2